+++ src/petro/extractor/avc_sample_extractor.cpp
@@ -61,10 +61,11 @@ std::string avc_sample_extractor::file_path() const
     return m_impl->file_path();
 }
 
-uint64_t avc_sample_extractor::timestamp() const
+/// Return the total media duration in microseconds
+uint64_t avc_sample_extractor::media_duration() const
 {
     assert(m_impl);
-    return m_impl->timestamp();
+    return m_impl->media_duration();
 }
 
 uint64_t avc_sample_extractor::decoding_timestamp() const